Chandru: Paging, Split the CSS files, Bugs in Tagging & listing, Decided Area won't be shown for pro and rest, only for shared

+++ zasz.me/Areas/Shared/Models/Post.cs
@@ -71,6 +71,6 @@ namespace zasz.me.Areas.Shared.Models
 
         List<Post> Page(int PageNumber, int PageSize, Site ProOrRest);
 
-        long Count(Site ProOrRest);
+        int Count(Site ProOrRest);
     }
 }
